refactor(TemplateBuilder): group PipeRenderer, PipeTypeSearchMenu, and PipeList into module directories
