+++ src/components/QueryEditor/TemplateBuilder/PipeList/PipeList.tsx
@@ -3,9 +3,9 @@ import React, { ReactNode } from 'react';
 import { useStyles2 } from '@grafana/ui';
 
 import { InsertableSeparator } from './InsertableSeparator';
-import { STREAM_TEMPLATE_TYPE } from './segmentHelpers';
-import { getStyles } from './styles';
-import { Pipe } from './types';
+import { STREAM_TEMPLATE_TYPE } from '../segmentHelpers';
+import { getStyles } from '../styles';
+import { Pipe } from '../types';
 
 interface Props {
   pipes: Pipe[];
