+++ src/components/QueryEditor/TemplateBuilder/PipeList/index.ts
@@ -1,0 +1,1 @@
+export { PipeList } from './PipeList';
