+++ src/components/QueryEditor/TemplateBuilder/PipeRenderer/PipeRenderer.tsx
@@ -4,9 +4,9 @@ import { IconButton, useStyles2 } from '@grafana/ui';
 
 import { OptionalExtensionMenu } from './OptionalExtensionMenu';
 import { SegmentRenderer } from './SegmentRenderer';
-import { getStyles } from './styles';
-import { getTemplate } from './templates/registry';
-import { Pipe } from './types';
+import { getStyles } from '../styles';
+import { getTemplate } from '../templates/registry';
+import { Pipe } from '../types';
 
 interface Props {
   pipe: Pipe;
