+++ src/components/QueryEditor/TemplateBuilder/PipeRenderer/index.ts
@@ -1,0 +1,2 @@
+export { PipeRenderer } from './PipeRenderer';
+export { PipeFieldLoadersProvider } from './PipeFieldLoadersProvider';
