+++ src/components/QueryEditor/TemplateBuilder/PipeTypeSearchMenu/index.ts
@@ -1,0 +1,1 @@
+export { PipeTypeSearchMenu } from './PipeTypeSearchMenu';
